drivers: safety-ivc: Fix Coverity defects
CID 495719, CID 495776 -
return of wait_event_interruptible_timeout is long.
Converted the timeout variable type to same.

CID 495762 -
String literal is considered a const array of chars.
Changed the function parameter to const char*

CID 495755 -
Breaking out of for loop if id matches. This disables
use of temp after freeing memory.

CID 495723, CID 495705, CID 495736 -
Converted dev_major to unsigned int as
unregister_chrdev_region uses unsigned input.

CID 495724 -
Freeing the req pointer in case of error to avoid
memory leak.

Bug 3461002

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2675798
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/platform/tegra/tegra_l1ss_cmd_resp_l2_interface.c
@@ -142,7 +142,7 @@ int user_send_service_status_notification(const nv_guard_srv_status_t *Var1,
 	//NvGuard_ReturnType_t lRet = NVGUARD_E_NOK;
 	uint8_t lBytePos = 0U;
 	const uint8_t *lPtr = (const uint8_t *)(Var1);
-	unsigned long timeout;
+	long timeout;
 
 	PDEBUG("SrvId = %d Status=%d ErrorInfoSize=%d ErrorInfo=%s\n",
 			Var1->srv_id, Var1->status,
@@ -207,7 +207,7 @@ int user_send_phase_notify(struct l1ss_data *ldata, nv_guard_3lss_layer_t layer,
 		nv_guard_tegraphase_t phase)
 {
 	static cmdresp_frame_ex_t send_phase = {0};
-	int timeout =
+	long timeout =
 		wait_event_interruptible_timeout(ldata->cmd.notify_waitq,
 						 (atomic_read(
 						 &ldata->cmd.notify_registered)
